Camera Sensor
I have added the new CameraSensor and did some changes on the original
Sensor and GBLSensor.

There is some additionnal code not used for the moment. I will use it in
the next commit (octree filtering, uncertainty projection).

The CameraSensor is a single camera working as a depth sensor (like
kinect).