Enable m_selectByTypeAndName even when there is only one child
This makes the menu entry to show up also on the root element of the
DBRoot tree. There is no need to require that the selected entry has at
least two children, while it still make sense to disable it for
leafObjects.

Also add a separator because this "selection" functionality is a new
distinct tool and it deserves it's own visibility.