Allow both strict and non-strict type match in selectByTypeAndName()
Add a new checkbox to choose between strict and non-strict type
selection.

This allows for instance to perform a strict match on CC_TYPES::MESH
which is particularly useful when selecting actual meshes with the
intent to merge them.

A non exclusive match for meshes would select also entities _containing_
meshes along with the actual mesh entities, which can be
counter-intuitive when asking to select meshes.