Fix calculating the dip direction of facets with normals pointing downwards
The concept of "dip direction" can be visualized like this:

  The dip direction is the direction (measured as a clockwise angle
  relative to the North) in which a liquid flows when poured on a plane
  from the top.

From this definition it is clear that the dip direction must be the same
for parallel facets, regardless of whether their normals point upwards
or downwards.

The formula using atan2() with the swapped N.x and N.y already gives the
correct results for facets with the normal pointing upwards, so just use
the sign of N.z to invert the normals if they point downwards.

Fixes #241