Simplify the calculation of the dip angle
The dip angle is the inclination a facet regardless of whether its
normal is pointing upwards or downwards: it's the angle between the
facet and the horizontal plane, measured clockwise with values in
[0,pi/2].

It can be calculated using the classic formulas for spherical
coordinates as the angle between an (x,y,z) vector of modulo r and the
zenith:

  phi = acos(z / r)

For planes with the normal pointing upwards, the formula above gives
exactly the dip angle (the angle between the plane and the horizon)
because the normal and the plane are perpendicular and so are the zenith
and the horizon.

Considering that the normal is a unit vector, and that for the dip we
have to ignore the orientation of the normal, we get this formula:

  dip = acos(fabs(N.z))

The sign could also be used, as in:

  dip = acos(sign(N.z) * N.z);

but using fabs() seems more natural and should also be faster than using
the multiplication.

Compared to the old formula this one using acos() is easier to
understand as it resemble the mathematical model more closely, and it's
also less expensive computationally, it gives the same results as the
old one with a maximum difference of +-3.8147e-06 degrees which is quite
small.