Changes some default paths
There are three paths we are interested in:

1) "next to" the app (next to the exe on Window, beside the .app on macOS)
2) place to store user data (file dialog defaults)
3) place to find plugins and other internal support files

This fixes (1) for macOS so it can find the global shift file.  It does this by setting the current directory on startup and using it to look for things.  So on the Mac it get set to beside the .app, on Windows it stays next to the exe.

This commit also changes dialogs to use the user's documents directory as the default for (2) by using QStandardPaths.

 (3) wasn't changed.