{macOS} Fix problem with the main window jumping around and resizing when dragging toolbars
It seems that setting the baseSize on the QMainWindow causes this flickering and resizing issue.  The docs imply that the baseSize is only used if sizeIncrement() is non-zero but that's obviously not the case.

"The base size is used to calculate a proper widget size if the widget defines sizeIncrement()."

It also implies in the docs for sizeIncrement() that this is ignored on Windows and possibly on X11 anyways, so I don't see any reason to set it to begin with.

Fixes #372