Use system locale to format number of points in several places
I seem to recall we went back-and-forth on some QLocale problems and I see that QLocale(QLocale::English) is used in several paces.

I've used QLocale::system() - if it's a problem we can switch this to QLocale::English, but it would be better if we could handle it properly for everyone instead of forcing English.

Fixes #559