{plugins} Allow overriding plugins by file name
Uses the search paths - any plugin with the same file name in a later path will replace the previous plugin.

Something like this was intended - as indicated by the comments - but it was not implemented properly. When implementing this I chose to use the file name, not the name inside the plugin.