On the command line - if a mesh can't be merged it would crash in the error message
There is a check for nullptr and then an immediate use of the pointer in the log message.

The fix for this lead to CLEntityDesc which was not being filled in properly in some default cases - "Unnamed cloud" & "Unnamed mesh". These would result in no names in the description because the string was being passed to QFile() in the CLEntityDesc constructor.