Better error reporting from E57
The e57::ImageFile constructor actually opens and possibly reads the file (!), so it's possible that the constructor will throw an exception. Move the constructors into the try {} so we get non-generic error messages.

Also use the e57 exception context (if it exists) to give move information and make error messages more non-dev-friendly.