Add another example - FindNearestCandidate() in ccCountourExtractor.cpp
This one is a little different because the for loop uses "continue" which won't work inside a labmda.

I tried keeping the same code for each by using returns instead of continue and converting
the "non-TBB" to using a lambda like this:

for (unsigned i=0; i<pointCount; ++i) [&]
{
...
} ();

But profiling showed this to be slightly slower in the non-TBB case.

So instead the entire for loop must be duplicated.

Speedups for extracting contours using the TBB version were 2-4x (I have 4 cores) across 4 sample models ranging from 155k-171M points.