Remove overrides and clean up virtuals in PointCloudTpl.h (#788)
As discussed in Slack, this template should not be overriding anything in the base class, so remove all override keywords. In addition, prune back the number of virtual functions to ones that are used.

This is kind of a fragile class since it's easy to unintentionally override or shadow members and functions in the base class, or mask the ones here in a class that inherits from this template.

One issue remains: clear() is in classes at all three levels, so it is not working as intended. This will require some discussion.