Use an enum for CC_TYPES/CC_CLASS_ENUM and use a namespace instead of a class
Now that we can specify enum types in C++, this is much cleaner.

It also avoids linking issues if you want to use these in a container (like a map). They are
declared and can be used as integral constant expressions (like switches), but they aren't
defined in the cpp, so they can't be used in a map (at least with clang).