{macOS} Fix launching CC when double-clicking application
On macOS, when double-clicking the application the Finder adds a command-line parameter like "-psn_0_582385" which is a "process serial number".

This caused the new check for "command-line mode" to be incorrect when launched as a GUI.

Instead, check for a terminal file descriptor for stdin.

This check (or variant of it) would probably work on other systems as well, but I've made it macOS-specific.