Point list picking: Avoid many, many unnecessary allocations
The table was creating a QTableWidgetItem for every cell in the table every time a new point was added. While it didn't leak memory because they were all cleaned up properly, it is unnecessary and was noticeable when picking many points.

Instead, figure out how many new rows we have, allocate the items for them, and just set the text inside the loop.

(Also fix a bunch of signed/unsigned warnings.)