{macOS} Statically link ffmpeg
Homebrew changed their policies of allowing options in their builds, so ffmpeg now has about 40 libraries it pulls in which makes it useless for shipping anything.

Normally I would wait for 2.11 to make this change, but it's too difficult to go back-and forth between versions.

I now build ffmpeg myself using the following options and link to it statically:

./configure \
--enable-gpl --enable-version3 \
--enable-pthreads \
--enable-hardcoded-tables \
--disable-all \
--disable-doc --disable-network --disable-sdl2 \
--enable-avcodec --enable-avformat --enable-swscale \
--enable-encoder=libx264 \
--enable-protocol=file \
--enable-muxer=mp4 \
--enable-libx264