Fix ODR violation: "Edge"
"Edge" is defined in three places - once as a class, twice as a struct. This violates the "One Definition Rule"

https://en.wikipedia.org/wiki/One_Definition_Rule

The result was that my debug build was getting the wrong one and crashing...

The fix is to put classes defined in cpp files into an anonymous namespace. This makes the class available only in that compilation unit.