Add "priority" to FileIOFilter::FilterInfo (#885)
* Add "priority" to FileIOFilter::FilterInfo

The order of the menu items used to be determined by the order FileIOFilters were registered. This meant
that plugins couldn't set the order and items would end up in different places depending on the plugin load
order. This order is now controlled by the priority so it is deterministic.

- any items with the same priority will be sorted by id
- also changed plugin-based filter ids from "+" to "_"

The order of the menu items should now be the same as it was before I started messing with FileIOFilters.

* Updated I/O filters priorities