Fix a disconnect that wasn't doing anything
When connections were changed to the new style, this disconnect was missed. It was also semi-broken before the change because it only disconnected one selection signal, not both.

Because we are using labmdas for these connections (required because arguments don't match), it's easiest to just not connect them then to track the connections for disconnect later.

(Also fixed error message to reflect the correct action.)