Track if we are running on the command line in ccApplicationBase
The plugin manager is going to need to know if we are running on the command line so we don't disable plugins in that case. Otherwise the user would need to run the GUI to enable/disable them.

Also rename ccApplicationBase::init() to ccApplicationBase::initOpenGL() to better reflect what it's doing and to remove inverted "!noOpenGLSupport" logic.