Fix assert in ccColorScale
Because the signals were hooked up first, the channels were being updated before the colour scales were updated.

This fired the assert(m_updated) in ccColorScale::getColorByRelativePos.

The solution is to construct everything, then hook up the signals, then update the channels.

This also save a lot of extra updating because the combo box would trigger an update for each channel for each SF.

Fixes #1136