Fixes plugins so they are recognized properly
The code in ccDefaultPluginInterface does in fact have to be compiled into each plugin.

To avoid relative paths, we set the path to it as a property on CCPluginAPI and then use that to add it to each plugin's sources.

We do not need to include this file on the application side.