{cmake} Update FBXIO plugin
I didn't change any of the linking and includes in the main CMakeLists file since I can't compile it.

If the SDK provides a proper cmake config file, you should be able to use find_package and then most of the additional FBX stuff in CMakeLists should not be required.