Add shared RDB library to install target
CMake already knows where the RDB library is because of the `rdbc`-target. Use
this information to add the library (and its symbolic links) to the install
target.

The installation of the rdb library can be controlled with the option
`PLUGIN_IO_QRDB_INSTALL_DEPENDENCY` and is `OFF` by default.