Add option PLUGIN_IO_QRDB_FETCH_DEPENDENCY to download RDB SDK (#1370)
* Add option PLUGIN_IO_QRDB_FETCH_DEPENDENCY to download RDB SDK

The latest RDB SDK can now be downloaded from
https://repository.riegl.com/software/libraries/rdblib/current

Add the option `PLUGIN_IO_QRDB_FETCH_DEPENDENCY` to optionally download and use
the SDK for building the QRDB IO plugin.

* Add shared RDB library to install target

CMake already knows where the RDB library is because of the `rdbc`-target. Use
this information to add the library (and its symbolic links) to the install
target.

The installation of the rdb library can be controlled with the option
`PLUGIN_IO_QRDB_INSTALL_DEPENDENCY` and is `OFF` by default.

* Enable PLUGIN_IO_QRDB plugin for CI builds

* RDBIO: replace hard coded 'lib' destination directory

* RDBIO: use _InstallFiles helper function for install

To install RDB runtime library into all the right folders use the helper
function `_InstallFiles` from `cmake/Install.cmake`.