Remove about dialog custom stylesheet
So that it respects the global stylesheet.

Related to #1415.
Commit #2ae7b490d9b07cdc2544b1dd117075cc9189f7ed forced the font color
to black, which indeed fixes the probleme of white text on white
background, but I think it is better the let the dialog use the 'global'
stylesheet so that if the theme is dark, the about window also is.