Linux: Fix shared lib install folder
On Linux the `INSTALL_RPATH` was set to
`"${CMAKE_INSTALL_FULL_LIBDIR}/cloudcompare`.

However the path where the shared libs were installed
was set to `"${destination}/${CMAKE_INSTALL_LIBDIR}/` or
`"${CMAKE_INSTALL_LIBDIR}"` which is not the same path as
the `INSTALL_RPATH`.

To illustrate, with a `CMAKE_INSTALL_PREFIX` set to
`~/CloudCompare/install` the shared libs were installed
in `~/CloudCompare/install/lib64` but the RPATH set in
each lib was pointing to `~/CloudCompare/install/lib64/cloudcompare`

Resulting in an error like this when running the installed CloudCompare:
```
error while loading shared libraries: libCCAppCommond.so: cannot open shared object file: No such file or directory
```

This fixes the problem by making the two things point to the same
path `"${CMAKE_INSTALL_FULL_LIBDIR}/cloudcompare"`.
So in the example libs properly end up in
`~/CloudCompare/install/lib64/cloudcompare`.

Although the cmake for installing shared target is still a bit ?strange?.