LAS: Allow entering custom scale values
Clicking on the "custom scale" radio button did not enable the
spin box, so the user could not enter the custom scale it wanted to use.

This commits adds 2 spin box, so that there a 3 (one per dimension X, Y, Z).
Also this commits makes it so that when clicking on the
custom scale radio button the 3 spin box are enabled and value can be changed,
when the radio button is no longer selected, the 3 spin box are disabled