Make a new suboption (MAX_NUMBER_OF_POINTS) for octree subsampling. (Refactoring as well)
it will try to guess an octree number based on number of points, but it will run iteratively, till it overflows the given number of points, so it will always result in the highest possible OCTREE number.

it will be slower then a given octree number, but it will be way faster than
load->subsample->save
load->subsample->save
from command line.

It won't give back a result exactly the given number of points as it does the RANDOM SAMPLING, but it will result a gridded cloud, which could make better understanding in downstream CAD applications.