Substantially improve speed of cloud merge operation (#2002)
* ccScalarField: use min/max getters

Allows for future changes to min/max storage and calculation.

* doActionMerge: computeMinAndMax only after all appends

Eliminates quadratic runtime behavior with large numbers of clouds.

* doActionMerge: reserve enough space for final cloud at start

Eliminates quadratic runtime behavior with large numbers of clouds.

* ccScalarField: speed up computeMinAndMax

The unsigned cast will already correctly floor, so we can remove the
call to floor and the double cast. Reduces computeMinAndMax execution
time by 45%.

* ccPointCloud: speed up append

The float->double and double->float cast, plus the addition, are a
measurable part of the execution time and often not needed.