LAS: assign by default scalar fields to extra LAS dim
Currently when saving a LAS file, scalar fields that are not mapped
to LAS standard field or that are not mapped to LAS extra field
(only possible if the cloud was read from a LAS), are not going to be
saved unless the user manually assigns CC scalar fields to LAS extra
fields, which can be tedious.

These changes makes CloudCompare automatically assigns as LAS extra
dimension any scalar fields that is not already assigned to standard
or extra LAS dimension using default settings (name == sf Name and type
== f32).

The "workflow" is that if the "Assign other fields as extra scalarfields"
checkbox of the 2nd tab is checked (which it is by default),
then the dialog will automatically assign to a extra LAS field with
default settings any CC scalar field that is not already assigned something,
this is done by creating a "card" in the 3rd tab.

Everytime the point format or the standard field mapping is changed, the
automatic assigning is done/re-done. Extra scalar field cards that were
automatically (default) made are "locked" as they may be automatically
removed/added when point format/standard field mapping changes, a user
can unlock the card to allow customizing it, in that case, the card is
no longer considerer a default.