Add "Don't ask again" option to quit confirmation
Some users do not like the dialog asking to confirm if
it really wanted to quit the app.

This adds a "Yes, don't ask again" option that will quit the app
and make CC remember not to show that confirmation dialog and
directly exit.

This setting can be toggled in the settings menu (Display->display
settings->Other Options).