Test and reset VBOs at the beginning of the LoD rendering
Prevent the simultaneous use of LoD and the retention of VBOs in VRAM.
Before the LoD construction, the point cloud is rendered using the "regular" rendering loop, resulting in the creation and filling of VBOs (if VBO rendering is enabled). After LoD construction, the VBO memory is not cleared. This PR use ReleaseVBO to test and reset VBO at the begining ot the LoD rendering.