LoD improvements (#2114)
* Remove LODAutoDisable members in ccGLWindowInterface

It seems to have no utility in practice.
it produces an invisible glitch by bypassing the LOD loop and entering the standard rendering loop. As a result, if VBOs are enabled, they are filled (if needed) and updated.

* Test and reset VBOs at the beginning of the LoD rendering

Prevent the simultaneous use of LoD and the retention of VBOs in VRAM.
Before the LoD construction, the point cloud is rendered using the "regular" rendering loop, resulting in the creation and filling of VBOs (if VBO rendering is enabled). After LoD construction, the VBO memory is not cleared. This PR use ReleaseVBO to test and reset VBO at the begining ot the LoD rendering.

* Update CHANGELOG

* minor header fixes

* Proposal: expose LOD parameters to PropertiesTree

* Review from @dgirardeau

* second round of review from @dgirardeau

Co-authored-by: Daniel Girardeau-Montaut <daniel.girardeau@gmail.com>

---------

Co-authored-by: Daniel Girardeau-Montaut <daniel.girardeau@gmail.com>