LAS: Fix OverlapFlag value read
The overlap flag is either 0 or 1, it is stored as a bit field.
By mistake we kept the value of the bit align in the bit field when
assigning it to a scalar field, meaning the value was either 0 or 8.

This fixes that so the value is a proper 0 or 1

Writing had no problem however