Las fixes (#2137)
* LAS: Fix OverlapFlag value read

The overlap flag is either 0 or 1, it is stored as a bit field.
By mistake we kept the value of the bit align in the bit field when
assigning it to a scalar field, meaning the value was either 0 or 8.

This fixes that so the value is a proper 0 or 1

Writing had no problem however

* LAS fix withheld flag association

A Small code mistake me use write the whithheld flag as the key point flag

* LAS: Sync extended_classification_flags with legacy

LASZip handles some backward compatibility for some flag values
which can be set at 2 different places when the file has the extended
point format.

However it requires us to have them synced otherwise it gives an
error