Don't omit empty fields
Previously, an empty CSV field would cause columns in a CSV to get misaligned.
E.g.
```csv
x,y
,42
```
would be the same as:
```csv
x,y
42
```

That is, "42" would be treated as if it were in the "x" column, not the "y" column. Now, empty fields are not discarded.