Fix duplicate assignment of LAS extra scalar fields
When saving a LAS file that originally had extra scalar fields
we could end up creating duplicates of them when saving.

This bug was due to the dialog auto assigning non standard field name to
new scalar extra scalar field with some default options. This was done
every time the point format changed (but here it was properly handled).

What was missing is that when we tell the dialog, which extra scalar
field we know of thanks to the meta data attached to the cloud, we did
not clear the default created extra scalar fields, so we ended up with
duplicates.