Restrict Windows CI to QT 6.8.3
Due to a change in the qt6-main package on conda-forge (version 6.9.3_1), our CI currently fails on Windows.
For now, we are restricting the CI to use the Qt 6.8.3 package, since it’s the version we use internally to build the Windows release.
This issue requires further investigation, as we may be doing something non-portable in the DeployQt script.