Fix Translation fusion process (for Qt6 and more broadly...)
Translation function actually does nothing because it relies on Qt5
private variable... but in fact the function is broken for a long time,
maybe since its inception... the merge process was trying to merge only
one file. It results on complicated function that only do a copy under
the hood) and the search of Qt translation path was not robust wrt. the
different OSes and distributions of Qt. It mostly went unoticed because
on Windows OSes windeployqt do most of the job by copying Qt translation
files.

We fix this behavior by:
- Using qtpaths to robustly retrieve Qt translations path (Not found
  other way to do so... maybe use QT6_INSTALL_TRANSLATION varialble
  under CMake)
- Eventually Merging something : the function should merge
  qtbase_<lang>.qm and qt_<lang.qm>