Pixi/macOS: run CC without install step, remove custom bundle script
Now that install step do a full analysis and copy in the bundle (.app)
of runtime dependencies (thanks to macdeployqt), the cmake install step
has a prohibitive runtime for simple debugging / testing loops. In this
commit, we changed the behavior of 'pixi run CloudCompare' to run
CloudCompare from the build tree rather from the installed package.
'Pixi run install' still pack the bundle.

We also removed the libBundleCloudCompare.py Python script that is now
useless but we kept signatureCloudCompare.py for the signature of the
bundle.

Not that we need to adapt CloudCompare-PythonRuntime to work with the
new macdeployqt install step.